+++ dev/src/sdlinterface.h
@@ -33,12 +33,12 @@ struct UpDown
 
 extern const int2 &GetFinger(int i, bool delta);
 extern UpDown GetKS(const char *name);
-extern float GetKeyTime(const char *name, int on);
+extern double GetKeyTime(const char *name, int on);
 extern int2 GetKeyPos(const char *name, int on);
 extern float GetJoyAxis(int i);
 
-extern float SDLTime();
-extern float SDLDeltaTime();
+extern double SDLTime();
+extern double SDLDeltaTime();
 
 extern int SDLWheelDelta();
